\name{data_binary}
\alias{data_binary}
\docType{data}
\title{Binary data}
\description{Binary variables for eight people}
\format{data.frame: 8 objects, 10 variables}
\source{Kaufman, L., Rousseeuw, P.J. (1990), \emph{Finding groups in data: an introduction to cluster analysis}, Wiley, New York, p. 24.}
\examples{
#library(clusterSim)
#data(data_binary)
#cluster.Sim(as.matrix(data_binary), p=5, 2, 6, icq="S", 
#outputCsv="results5", distances=c("b1","b3","b6"))
}
\keyword{datasets}
