\name{data_ordinal}
\alias{data_ordinal}
\docType{data}
\title{Ordinal data}
\description{Artificial ordinal data}
\format{data.frame: 26 objects, 12 variables}
\source{Artificial data}
\examples{
#library(clusterSim)
#data(data_ordinal)
#cluster.Sim(data_ordinal, 3, 3, 12,"S", 
#outputCsv2="results3")
}
\keyword{datasets}
