\name{data_ratio}
\alias{data_ratio}
\docType{data}
\title{Ratio data}
\description{Artificially generated ratio data}
\format{data.frame: 75 objects, 5 variables, 5-class structure}
\source{Artificially generated data}
\examples{
#library(clusterSim)
#data(data_ratio)
#c <- pam(data_ratio,10)
#index.G1(data_ratio, c$clustering)
}
\keyword{datasets}
