\name{dist.Symbolic}
\alias{dist.Symbolic}
\title{Calculates distance beetwen symbolic interval data}
\description{dist.Symbolic}
\usage{dist.Symbolic(data,type="U_2",gamma=0.5,power=2)}
\arguments{
\item{data}{symbolic data}
\item{type}{U\_2, M, H, S}
\item{gamma}{}
\item{power}{}
}
\details{
See file \$R\_HOME\\library\\clusterSim\\pdf\\distSymbolic\_details.pdf for further details
}
\author{
Marek Walesiak \email{Marek.Walesiak@ae.jgora.pl}, Andrzej Dudek \email{Andrzej.Dudek@ae.jgora.pl} 

Department of Econometrics and Computer Science, University of Economics, Wroclaw, Poland \url{http://www.ae.jgora.pl/keii}
}
\references{

Billard, L., Diday, E. (2006): \emph{Symbolic data analysis. Conceptual statistics and data mining}, Wiley, Chichester.

}
\seealso{
\code{\link{dist.SM}}
}
\examples{
}
\keyword{cluster}
